%%  WARNING: This file was automatically generated from the associated 
%%  it_plot.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the it_plot.mid file in the project MID directory. Once the 
%%  it_plot.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the em function

\name{em}
\alias{em}
\title{Typesetting unit}
\concept{splus}
\usage{em()}
\description{Return the width and height of the letter 'm' in the current font.}


\value{
a two-element vector containing the width and height, respectively, of the letter 'm' in user coordinates for the current font.
}
\seealso{
\code{\link{stringSize}}.}

\examples{
em()
}
\keyword{utilities}

